+++ packages/backend-deployer/src/cdk_deployer.test.ts
@@ -1,7 +1,11 @@
 import { after, beforeEach, describe, it, mock } from 'node:test';
 import { CDKDeployer } from './cdk_deployer.js';
 import assert from 'node:assert';
-import { AmplifyError, BackendLocator } from '@aws-amplify/platform-core';
+import {
+  AmplifyError,
+  AmplifyUserError,
+  BackendLocator,
+} from '@aws-amplify/platform-core';
 import { DeployProps } from './cdk_deployer_singleton_factory.js';
 import { CDKDeploymentError, CdkErrorMapper } from './cdk_error_mapper.js';
 import {
@@ -43,23 +47,23 @@ void describe('invokeCDKCommand', () => {
     backendLocator,
     packageManagerControllerMock as never
   );
-  const execaMock = mock.method(invoker, 'executeCommand', () =>
+  const executeCommandMock = mock.method(invoker, 'executeCommand', () =>
     Promise.resolve()
   );
 
   beforeEach(() => {
-    execaMock.mock.resetCalls();
+    executeCommandMock.mock.resetCalls();
   });
 
   after(() => {
-    execaMock.mock.restore();
+    executeCommandMock.mock.restore();
   });
 
   void it('handles options for branch deployments', async () => {
     await invoker.deploy(branchBackendId);
-    assert.strictEqual(execaMock.mock.callCount(), 2);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 17);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 2);
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'synth',
       '--ci',
@@ -78,7 +82,7 @@ void describe('invokeCDKCommand', () => {
       'amplify-backend-type=branch',
       '--quiet',
     ]);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -100,9 +104,9 @@ void describe('invokeCDKCommand', () => {
 
   void it('handles deployProps for sandbox', async () => {
     await invoker.deploy(sandboxBackendId, sandboxDeployProps);
-    assert.strictEqual(execaMock.mock.callCount(), 2);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 19);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 2);
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 19);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'synth',
       '--ci',
@@ -125,8 +129,8 @@ void describe('invokeCDKCommand', () => {
       }`,
       '--quiet',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[0]?.length, 18);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 18);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -152,9 +156,9 @@ void describe('invokeCDKCommand', () => {
 
   void it('handles options and deployProps for sandbox', async () => {
     await invoker.deploy(sandboxBackendId, sandboxDeployProps);
-    assert.strictEqual(execaMock.mock.callCount(), 2);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 19);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 2);
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 19);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'synth',
       '--ci',
@@ -177,8 +181,8 @@ void describe('invokeCDKCommand', () => {
       }`,
       '--quiet',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[0]?.length, 18);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 18);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -204,9 +208,9 @@ void describe('invokeCDKCommand', () => {
 
   void it('handles destroy for sandbox', async () => {
     await invoker.destroy(sandboxBackendId);
-    assert.strictEqual(execaMock.mock.callCount(), 1);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 15);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 1);
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 15);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'destroy',
       '--ci',
@@ -229,9 +233,10 @@ void describe('invokeCDKCommand', () => {
     await invoker.deploy(branchBackendId, {
       validateAppSources: true,
     });
-    assert.strictEqual(execaMock.mock.callCount(), 4);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 17);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 4);
+
+    // Call 0 -> synth
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'synth',
       '--ci',
@@ -250,23 +255,30 @@ void describe('invokeCDKCommand', () => {
       `amplify-backend-type=branch`,
       '--quiet',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[0]?.length, 4);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+
+    // Call 1 -> tsc showConfig
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 4);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'tsc',
       '--showConfig',
       '--project',
       'amplify',
     ]);
-    assert.equal(execaMock.mock.calls[2].arguments[0]?.length, 5);
-    assert.deepStrictEqual(execaMock.mock.calls[2].arguments[0], [
+
+    // Call 2 -> tsc
+    assert.equal(executeCommandMock.mock.calls[2].arguments[0]?.length, 5);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[2].arguments[0], [
       'tsc',
       '--noEmit',
       '--skipLibCheck',
       '--project',
       'amplify',
     ]);
-    assert.equal(execaMock.mock.calls[3].arguments[0]?.length, 16);
-    assert.deepStrictEqual(execaMock.mock.calls[3].arguments[0], [
+
+    // Call 3 -> deploy
+    assert.equal(executeCommandMock.mock.calls[3].arguments[0]?.length, 16);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[3].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -290,9 +302,11 @@ void describe('invokeCDKCommand', () => {
     await invoker.deploy(sandboxBackendId, {
       validateAppSources: true,
     });
-    assert.strictEqual(execaMock.mock.callCount(), 4);
-    assert.equal(execaMock.mock.calls[0].arguments[0]?.length, 17);
-    assert.deepStrictEqual(execaMock.mock.calls[0].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 4);
+
+    // Call 0 -> synth
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
       'cdk',
       'synth',
       '--ci',
@@ -311,23 +325,29 @@ void describe('invokeCDKCommand', () => {
       '--method=direct',
       '--quiet',
     ]);
-    assert.equal(execaMock.mock.calls[1].arguments[0]?.length, 4);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+
+    // Call 1 -> tsc showConfig
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 4);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'tsc',
       '--showConfig',
       '--project',
       'amplify',
     ]);
-    assert.equal(execaMock.mock.calls[2].arguments[0]?.length, 5);
-    assert.deepStrictEqual(execaMock.mock.calls[2].arguments[0], [
+
+    // Call 2 -> tsc
+    assert.equal(executeCommandMock.mock.calls[2].arguments[0]?.length, 5);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[2].arguments[0], [
       'tsc',
       '--noEmit',
       '--skipLibCheck',
       '--project',
       'amplify',
     ]);
-    assert.equal(execaMock.mock.calls[3].arguments[0]?.length, 16);
-    assert.deepStrictEqual(execaMock.mock.calls[3].arguments[0], [
+
+    // Call 3 -> deploy
+    assert.equal(executeCommandMock.mock.calls[3].arguments[0]?.length, 16);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[3].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -349,7 +369,7 @@ void describe('invokeCDKCommand', () => {
 
   void it('disables type checking when tsconfig is not present', async () => {
     // simulate first execa call as throwing error when checking for tsconfig.json
-    execaMock.mock.mockImplementation((commandArgs: string[]) => {
+    executeCommandMock.mock.mockImplementation((commandArgs: string[]) => {
       if (commandArgs.includes('tsc')) {
         return Promise.reject(new Error('some error'));
       }
@@ -358,16 +378,42 @@ void describe('invokeCDKCommand', () => {
     await invoker.deploy(branchBackendId, {
       validateAppSources: true,
     });
-    assert.strictEqual(execaMock.mock.callCount(), 3);
-    assert.equal(execaMock.mock.calls[1].arguments[0]?.length, 4);
-    assert.deepStrictEqual(execaMock.mock.calls[1].arguments[0], [
+    assert.strictEqual(executeCommandMock.mock.callCount(), 3);
+
+    // Call 0 -> synth
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
+      'cdk',
+      'synth',
+      '--ci',
+      '--app',
+      "'npx tsx amplify/backend.ts'",
+      '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
+      '--context',
+      `amplify-backend-namespace=123`,
+      '--context',
+      `amplify-backend-name=testBranch`,
+      '--require-approval',
+      'never',
+      '--context',
+      `amplify-backend-type=branch`,
+      '--quiet',
+    ]);
+
+    // Call 1 -> tsc showConfig
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 4);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
       'tsc',
       '--showConfig',
       '--project',
       'amplify',
     ]);
-    assert.equal(execaMock.mock.calls[2].arguments[0]?.length, 16);
-    assert.deepStrictEqual(execaMock.mock.calls[2].arguments[0], [
+
+    // Call 2 -> Deploy (Skipping tsc)
+    assert.equal(executeCommandMock.mock.calls[2].arguments[0]?.length, 16);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[2].arguments[0], [
       'cdk',
       'deploy',
       '--ci',
@@ -384,6 +430,150 @@ void describe('invokeCDKCommand', () => {
       'never',
       '--context',
       'amplify-backend-type=branch',
+    ]);
+  });
+
+  void it('run typescript even if synth fails and throws TS error', async () => {
+    // simulate first execa call for synth as throwing error
+    executeCommandMock.mock.mockImplementation((commandArgs: string[]) => {
+      if (commandArgs.includes('synth')) {
+        throw new Error('some synth error');
+      }
+      if (commandArgs.includes('tsc') && commandArgs.includes('--noEmit')) {
+        throw new Error('some tsc error');
+      }
+      return;
+    });
+
+    await assert.rejects(
+      () =>
+        invoker.deploy(sandboxBackendId, {
+          validateAppSources: true,
+        }),
+      new AmplifyUserError(
+        'SyntaxError',
+        {
+          message: 'TypeScript validation check failed.',
+          resolution:
+            'Fix the syntax and type errors in your backend definition.',
+        },
+        new Error('some tsc error')
+      )
+    );
+    assert.strictEqual(executeCommandMock.mock.callCount(), 3);
+
+    // Call 0 -> synth
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
+      'cdk',
+      'synth',
+      '--ci',
+      '--app',
+      "'npx tsx amplify/backend.ts'",
+      '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
+      '--context',
+      'amplify-backend-namespace=foo',
+      '--context',
+      'amplify-backend-name=bar',
+      '--context',
+      'amplify-backend-type=sandbox',
+      '--hotswap-fallback',
+      '--method=direct',
+      '--quiet',
+    ]);
+
+    // Call 1 -> tsc showConfig (ts checks are still run)
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 4);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
+      'tsc',
+      '--showConfig',
+      '--project',
+      'amplify',
+    ]);
+
+    // Call 2 -> tsc
+    assert.equal(executeCommandMock.mock.calls[2].arguments[0]?.length, 5);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[2].arguments[0], [
+      'tsc',
+      '--noEmit',
+      '--skipLibCheck',
+      '--project',
+      'amplify',
+    ]);
+  });
+
+  void it('throws the original synth error if the synth failed but tsc succeeded', async () => {
+    // simulate first execa call for synth as throwing error
+    const stderr = `some rubbish before
+Error: some cdk synth error
+    at lookup (/some_random/path.js:1:3005)
+    at lookup2 (/some_random/path2.js:2:3005)`;
+    executeCommandMock.mock.mockImplementation((commandArgs: string[]) => {
+      if (commandArgs.includes('synth')) {
+        return Promise.reject(new Error(stderr));
+      }
+      return Promise.resolve();
+    });
+
+    await assert.rejects(
+      () =>
+        invoker.deploy(sandboxBackendId, {
+          validateAppSources: true,
+        }),
+      new AmplifyUserError(
+        'BackendSynthError',
+        {
+          message: 'Unable to build the Amplify backend definition.',
+          resolution:
+            'Check your backend definition in the `amplify` folder for syntax and type errors.',
+        },
+        new Error(`Error: some cdk synth error
+    at lookup (/some_random/path.js:1:3005)`)
+      )
+    );
+    assert.strictEqual(executeCommandMock.mock.callCount(), 3);
+
+    // Call 0 -> synth
+    assert.equal(executeCommandMock.mock.calls[0].arguments[0]?.length, 17);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[0].arguments[0], [
+      'cdk',
+      'synth',
+      '--ci',
+      '--app',
+      "'npx tsx amplify/backend.ts'",
+      '--all',
+      '--output',
+      '.amplify/artifacts/cdk.out',
+      '--context',
+      'amplify-backend-namespace=foo',
+      '--context',
+      'amplify-backend-name=bar',
+      '--context',
+      'amplify-backend-type=sandbox',
+      '--hotswap-fallback',
+      '--method=direct',
+      '--quiet',
+    ]);
+
+    // Call 1 -> tsc showConfig (ts checks are still run)
+    assert.equal(executeCommandMock.mock.calls[1].arguments[0]?.length, 4);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[1].arguments[0], [
+      'tsc',
+      '--showConfig',
+      '--project',
+      'amplify',
+    ]);
+
+    // Call 2 -> tsc
+    assert.equal(executeCommandMock.mock.calls[2].arguments[0]?.length, 5);
+    assert.deepStrictEqual(executeCommandMock.mock.calls[2].arguments[0], [
+      'tsc',
+      '--noEmit',
+      '--skipLibCheck',
+      '--project',
+      'amplify',
     ]);
   });