+++ packages/cli/src/commands/sandbox/sandbox-delete/sandbox_delete_command.ts
@@ -1,12 +1,14 @@
 import { ArgumentsCamelCase, Argv, CommandModule } from 'yargs';
 import { SandboxSingletonFactory } from '@aws-amplify/sandbox';
 import { AmplifyPrompter } from '@aws-amplify/cli-core';
+import { SandboxCommandGlobalOptions } from '../option_types.js';
+import { ArgumentsKebabCase } from '../../../kebab_case.js';
 
 /**
  * Command that deletes the sandbox environment.
  */
 export class SandboxDeleteCommand
-  implements CommandModule<object, SandboxDeleteCommandOptions>
+  implements CommandModule<object, SandboxDeleteCommandOptionsKebabCase>
 {
   /**
    * @inheritDoc
@@ -30,7 +32,7 @@ export class SandboxDeleteCommand
    * @inheritDoc
    */
   handler = async (
-    args: ArgumentsCamelCase<SandboxDeleteCommandOptions>
+    args: ArgumentsCamelCase<SandboxDeleteCommandOptionsKebabCase>
   ): Promise<void> => {
     let isConfirmed = args.yes;
     if (!isConfirmed) {
@@ -50,36 +52,19 @@ export class SandboxDeleteCommand
   /**
    * @inheritDoc
    */
-  builder = (yargs: Argv): Argv<SandboxDeleteCommandOptions> => {
-    return yargs
-      .option('yes', {
-        describe:
-          'Do not ask for confirmation before deleting the sandbox environment',
-        type: 'boolean',
-        array: false,
-        alias: 'y',
-      })
-      .option('name', {
-        describe:
-          'An optional name to distinguish between different sandbox environments. Default is the name in your package.json',
-        type: 'string',
-        array: false,
-      })
-      .check((argv) => {
-        if (argv.name) {
-          const projectNameRegex = /^[a-zA-Z0-9-]{1,15}$/;
-          if (!argv.name.match(projectNameRegex)) {
-            throw new Error(
-              `--name should match [a-zA-Z0-9-] and less than 15 characters.`
-            );
-          }
-        }
-        return true;
-      });
+  builder = (yargs: Argv): Argv<SandboxDeleteCommandOptionsKebabCase> => {
+    return yargs.option('yes', {
+      describe:
+        'Do not ask for confirmation before deleting the sandbox environment',
+      type: 'boolean',
+      array: false,
+      alias: 'y',
+    });
   };
 }
 
-export type SandboxDeleteCommandOptions = {
-  yes?: boolean;
-  name?: string;
-};
+type SandboxDeleteCommandOptionsKebabCase = ArgumentsKebabCase<
+  {
+    yes?: boolean;
+  } & SandboxCommandGlobalOptions
+>;