+++ packages/create-amplify/templates/basic-auth-data/amplify/data/resource.ts
@@ -53,6 +53,6 @@ Fetch records from the database and use them in your frontend component.
 
 /* For example, in a React component, you can use this snippet in your
   function's RETURN statement */
-// const { data: todos } = client.models.Todo.list()
+// const { data: todos } = await client.models.Todo.list()
 
 // return <ul>{todos.map(todo => <li key={todo.id}>{todo.content}</li>)}</ul>