+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/amplify/storage/resource.ts
@@ -13,7 +13,7 @@ export const storage = defineStorage({
     onUpload,
   },
   access: (allow) => ({
-    '/public/*': [
+    'public/*': [
       allow.resource(defaultNodeFunc).to(['read', 'write']),
       allow.resource(node16Func).to(['read', 'write']),
     ],