+++ scripts/components/api-changes-validator/test-resources/test-projects/without-breaks/project-with-namespace/src/index.ts
@@ -1,0 +1,3 @@
+import * as someNamespace from './some_namespace.js';
+
+export { someNamespace };