+++ packages/integration-tests/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @aws-amplify/integration-tests
 
+## 0.5.0-beta.5
+
+### Minor Changes
+
+- 0d1b00e: Update generated env package location and use the $ symbol
+
+## 0.5.0-beta.4
+
+### Patch Changes
+
+- 5e12247: feat(client-config): Generate client configuration based on a unified JSON schema
+
 ## 0.5.0-beta.3
 
 ### Patch Changes