+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @aws-amplify/sandbox
 
+## 0.5.2-beta.10
+
+### Patch Changes
+
+- Updated dependencies [8d73779]
+- Updated dependencies [0d1b00e]
+  - @aws-amplify/client-config@0.9.0-beta.8
+  - @aws-amplify/cli-core@0.5.0-beta.6
+
+## 0.5.2-beta.9
+
+### Patch Changes
+
+- Updated dependencies [7537216]
+- Updated dependencies [d95ab02]
+  - @aws-amplify/cli-core@0.5.0-beta.5
+  - @aws-amplify/client-config@0.9.0-beta.7
+
+## 0.5.2-beta.8
+
+### Patch Changes
+
+- Updated dependencies [4cd282e]
+  - @aws-amplify/client-config@0.9.0-beta.6
+
+## 0.5.2-beta.7
+
+### Patch Changes
+
+- Updated dependencies [b0112e3]
+- Updated dependencies [5e12247]
+  - @aws-amplify/deployed-backend-client@0.4.0-beta.4
+  - @aws-amplify/client-config@0.9.0-beta.5
+  - @aws-amplify/platform-core@0.5.0-beta.3
+  - @aws-amplify/backend-deployer@0.5.1-beta.4
+  - @aws-amplify/backend-secret@0.4.5-beta.3
+  - @aws-amplify/cli-core@0.5.0-beta.4
+
 ## 0.5.2-beta.6
 
 ### Patch Changes